third_party/WebKit/LayoutTests/screen_orientation/lockOrientation-bad-array-argument-expected.txt

Validates that lockOrientation() returns false on bad orientation value in array argument.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS screen.orientation.type is "portrait-primary"
PASS screen.lockOrientation(['portrait-primary', 'invalid-orientation']) threw exception TypeError: Failed to execute 'lockOrientation' on 'Screen': parameter 1 ('portrait-primary,invalid-orientation') is not a valid enum value..
PASS screen.lockOrientation(['portrait-primary', null]) threw exception TypeError: Failed to execute 'lockOrientation' on 'Screen': parameter 1 ('portrait-primary,') is not a valid enum value..
PASS screen.lockOrientation(['portrait-primary', undefined]) threw exception TypeError: Failed to execute 'lockOrientation' on 'Screen': parameter 1 ('portrait-primary,') is not a valid enum value..
PASS screen.lockOrientation(['portrait-primary', 123]) threw exception TypeError: Failed to execute 'lockOrientation' on 'Screen': parameter 1 ('portrait-primary,123') is not a valid enum value..
PASS screen.lockOrientation(['portrait-primary', window]) threw exception TypeError: Failed to execute 'lockOrientation' on 'Screen': parameter 1 ('portrait-primary,[object Window]') is not a valid enum value..
PASS screen.orientation.type is "portrait-primary"
PASS successfullyParsed is true

TEST COMPLETE
